refactor(auth): inline request header merge
Collapse the single-caller request header helper into packages/auth/src/create-oauth-app-auth.ts:232 so auth.fetch owns header merging and bearer mutation in one file. This deletes packages/auth/src/request-headers.ts after its only non-test caller was counted.
